refactor: move /ship flow into command module
